gcode_move: Move GCodeMove class from gcode.py to new extras module
