+++ klippy/toolhead.py
@@ -268,7 +268,8 @@ class ToolHead:
                                desc=self.cmd_SET_VELOCITY_LIMIT_help)
         gcode.register_command('M204', self.cmd_M204)
         # Load some default modules
-        modules = ["idle_timeout", "statistics", "manual_probe", "tuning_tower"]
+        modules = ["gcode_move", "idle_timeout", "statistics", "manual_probe",
+                   "tuning_tower"]
         for module_name in modules:
             self.printer.load_object(config, module_name)
     # Print time tracking
